+++ packages/core-js-compat/src/data.mjs
@@ -1416,7 +1416,10 @@ export const data = {
     chrome: '92',
     firefox: '90',
   },
+  // TODO: Remove from `core-js@4`
   'esnext.array.filter-out': {
+  },
+  'esnext.array.filter-reject': {
   },
   'esnext.array.find-last': {
   },
@@ -1669,7 +1672,10 @@ export const data = {
     chrome: '92',
     firefox: '90',
   },
+  // TODO: Remove from `core-js@4`
   'esnext.typed-array.filter-out': {
+  },
+  'esnext.typed-array.filter-reject': {
   },
   'esnext.typed-array.find-last': {
   },