+++ packages/core-js-pure/override/modules/esnext.typed-array.at.js
@@ -1,1 +1,1 @@
-// empty
+// TODO: Remove from `core-js@4`