+++ packages/core-js/features/array/filter-out.js
@@ -1,3 +1,4 @@
+// TODO: Remove from `core-js@4`
 require('../../modules/esnext.array.filter-out');
 var entryUnbind = require('../../internals/entry-unbind');