+++ packages/core-js/features/array/virtual/filter-out.js
@@ -1,3 +1,4 @@
+// TODO: Remove from `core-js@4`
 require('../../../modules/esnext.array.filter-out');
 var entryVirtual = require('../../../internals/entry-virtual');