+++ packages/core-js/modules/esnext.typed-array.filter-reject.js
@@ -1,5 +1,4 @@
 'use strict';
-// TODO: Remove from `core-js@4`
 var ArrayBufferViewCore = require('../internals/array-buffer-view-core');
 var $filterReject = require('../internals/array-iteration').filterReject;
 var fromSpeciesAndList = require('../internals/typed-array-from-species-and-list');
@@ -7,9 +6,9 @@ var fromSpeciesAndList = require('../internals/typed-array-from-species-and-list
 var aTypedArray = ArrayBufferViewCore.aTypedArray;
 var exportTypedArrayMethod = ArrayBufferViewCore.exportTypedArrayMethod;
 
-// `%TypedArray%.prototype.filterOut` method
+// `%TypedArray%.prototype.filterReject` method
 // https://github.com/tc39/proposal-array-filtering
-exportTypedArrayMethod('filterOut', function filterOut(callbackfn /* , thisArg */) {
+exportTypedArrayMethod('filterReject', function filterReject(callbackfn /* , thisArg */) {
   var list = $filterReject(aTypedArray(this), callbackfn, arguments.length > 1 ? arguments[1] : undefined);
   return fromSpeciesAndList(this, list);
 });